♻️ [refactor] separate into MainView and MainViewModel #21
