+++ Yeowoo/Sources/Models/MainModel/MainModel.swift
@@ -1,0 +1,16 @@
+//
+//  MainModel.swift
+//  Yeowoo
+//
+//  Created by 김용주 on 2023/07/20.
+//
+
+import SwiftUI
+
+struct Dummy: Identifiable, Hashable {
+    var id = UUID()
+    var albumName: String
+    var startDay: String
+    var endDay: String
+    var withPerson: [Color]
+}
