+++ Yeowoo/Sources/Views/Main/MainView.swift
@@ -8,11 +8,12 @@
 import SwiftUI
 
 struct MainView: View {
-    @State var traveling = 0
-    @State var openAlarm = false
-    @State var hasAlarm = true
-    @State var hasAlbum = false
-    @State var today: String = ""
+    @ObservedObject var mainViewModel = MainViewModel()
+//    @State var traveling = 0
+//    @State var openAlarm = false
+//    @State var hasAlarm = true
+//    @State var hasAlbum = false
+//    @State var today: String = ""
     
     var body: some View {
         NavigationView {
@@ -29,13 +30,13 @@ struct MainView: View {
                     
                     HStack {
                         NavigationLink(destination: EmptyView(),
-                                       isActive: $openAlarm) {
+                                       isActive: $mainViewModel.openAlarm) {
                             Button(action: {
-                                openAlarm.toggle()
-                                hasAlarm = false
+                                mainViewModel.openAlarm.toggle()
+                                mainViewModel.hasAlarm = false
                             }) {
                                 ZStack {
-                                    Image(hasAlarm == true ? "AlarmBell" : "Bell")
+                                    Image(mainViewModel.hasAlarm == true ? "AlarmBell" : "Bell")
                                         .resizable()
                                         .aspectRatio(contentMode: .fit)
                                         .frame(width: UIScreen.getHeight(48))
@@ -54,7 +55,6 @@ struct MainView: View {
                                     .frame(width: UIScreen.getHeight(48))
                             }
                         }
-                            
                     }
                     
                     Spacer()
@@ -62,14 +62,14 @@ struct MainView: View {
                 }
                 .frame(height: UIScreen.getHeight(48))
                 Spacer()
-                    .frame(height: UIScreen.getHeight(hasAlbum == true ? 35 : 84))
+                    .frame(height: UIScreen.getHeight(mainViewModel.hasAlbum == true ? 35 : 84))
                 
                 ZStack {
-                    if hasAlbum {
+                    if mainViewModel.hasAlbum {
                         ScrollView(.vertical, showsIndicators: true, content: {
                             VStack {
-                                if !(traveling == 2) {
-                                    if traveling == 1{
+                                if !(mainViewModel.traveling == 2) {
+                                    if mainViewModel.traveling == 1{
                                         RecommendLayout()
                                         Spacer()
                                             .frame(height: UIScreen.getHeight(41))
@@ -80,7 +80,7 @@ struct MainView: View {
                                     HStack {
                                         Spacer()
                                             .frame(width: UIScreen.getWidth(20))
-                                        if traveling == 1{
+                                        if mainViewModel.traveling == 1{
                                             Text("지금 쌓고 있는 추억")
                                                 .font(.custom18bold())
                                             Spacer()
@@ -94,7 +94,7 @@ struct MainView: View {
                                     
                                     Spacer()
                                         .frame(height: UIScreen.getHeight(45))
-                                } else if dummyData[dummyData.count-1].endDay == today {
+                                } else if dummyData[dummyData.count-1].endDay == mainViewModel.today {
                                     DoneTravelLayout(albumName: dummyData[dummyData.count-1].albumName)
                                     
                                     Spacer()
@@ -110,7 +110,7 @@ struct MainView: View {
                                     
                                     Spacer()
                                 }
-                                ForEach(dummyData[0..<(traveling == 2 ? dummyData.count : dummyData.count-1)].reversed(),
+                                ForEach(dummyData[0..<(mainViewModel.traveling == 2 ? dummyData.count : dummyData.count-1)].reversed(),
                                         id: \.self) { data in
                                     AlbumLayout(arr: data.withPerson, travelName: data.albumName, startDay: data.startDay, endDay: data.endDay)
                                         .padding(.bottom, UIScreen.getHeight(10))
@@ -128,40 +128,10 @@ struct MainView: View {
                         Spacer()
                         
                         HStack {
-//                            Spacer()
-//                                .frame(width: UIScreen.getWidth(20))
-//
-//                            ZStack {
-//                                RoundedRectangle(cornerRadius: 100)
-//                                    .foregroundColor(Color(uiColor: .secondarySystemBackground))
-//                                    .frame(width: UIScreen.getWidth(158), height: UIScreen.getHeight(76))
-//                                HStack {
-//                                    Button(action: {
-//                                        hasAlarm = false
-//                                    }) {
-//                                        Image(hasAlarm == true ? "hasAlarm" : "noAlarm")
-//                                            .resizable()
-//                                            .aspectRatio(contentMode: .fit)
-//                                            .frame(width: UIScreen.getHeight(48))
-//                                    }
-//
-//                                    Spacer()
-//                                        .frame(width: UIScreen.getWidth(16))
-//
-//                                    Button(action: {
-//
-//                                    }) {
-//                                        Image("Gear")
-//                                            .resizable()
-//                                            .aspectRatio(contentMode: .fit)
-//                                            .frame(width: UIScreen.getHeight(48))
-//                                    }
-//                                }
-//                            }
                             
                             Spacer()
                             
-                            if (traveling == 0 || traveling == 1) && hasAlbum {
+                            if (mainViewModel.traveling == 0 || mainViewModel.traveling == 1) && mainViewModel.hasAlbum {
                                 CameraButton()
                                     .padding(.trailing, UIScreen.getWidth(20))
                             } else {
@@ -180,9 +150,9 @@ struct MainView: View {
             }
         }
         .onAppear {
-            traveling = hasTraveling()
-            getCurrentDateTime(&today)
-            hasAlbum = hasEmpty()
+            mainViewModel.hasTraveling()
+            mainViewModel.getCurrentDateTime()
+            mainViewModel.hasEmpty()
         }
     }
 }
